+++ firmware/si_trx_defs.h
@@ -355,8 +355,8 @@ static uint8_t _si_trx_set_property_16(uint8_t group, uint8_t property, uint16_t
   buffer[1] = group;
   buffer[2] = 2;
   buffer[3] = property;
-  buffer[4] = (value >> 8);
-  buffer[5] = (value);
+  buffer[4] = (uint8_t)(value >> 8);
+  buffer[5] = (uint8_t)value;
 
   return _si_trx_transfer(6, 0, buffer);
 }
@@ -368,9 +368,9 @@ static uint8_t _si_trx_set_property_24(uint8_t group, uint8_t property, uint32_t
   buffer[1] = group;
   buffer[2] = 3;
   buffer[3] = property;
-  buffer[4] = (value >> 16);
-  buffer[5] = (value >> 8);
-  buffer[6] = (value);
+  buffer[4] = (uint8_t)(value >> 16);
+  buffer[5] = (uint8_t)(value >> 8);
+  buffer[6] = (uint8_t)value;
 
   return _si_trx_transfer(7, 0, buffer);
 }
@@ -382,10 +382,10 @@ static uint8_t _si_trx_set_property_32(uint8_t group, uint8_t property, uint32_t
   buffer[1] = group;
   buffer[2] = 4;
   buffer[3] = property;
-  buffer[4] = (value >> 24);
+  buffer[4] = (uint8_t)(value >> 24);
   buffer[5] = (value >> 16);
-  buffer[6] = (value >> 8);
-  buffer[7] = (value);
+  buffer[6] = (uint8_t)(value >> 8);
+  buffer[7] = (uint8_t)value;
 
   return _si_trx_transfer(8, 0, buffer);
 }
